support cascade standby
We support cascade standby now, a cascade standby will get WAL from
a standby asynchronously. To support flexible deployment, we add a
new GUC available_zone to identify the available zone of a cascade
standby, which will try to connect a standby in same available zone.

Users can use utility gs_ctl for failover and switchover. A cascade
standby will be promoted to be a standby when failover or switchover,
while a standby will be demoted to be a cascade standby after switchover.