合入PG补丁，从非扁平子查询中删除不必要的输出表达式。
PG提交号：55d5b3c08279b487cfa44d4b6e6eea67a0af89e4
Remove unnecessary output expressions from unflattened subqueries.

If a sub-select-in-FROM gets flattened into the upper query, then we
naturally get rid of any output columns that are defined in the sub-select
text but not actually used in the upper query.  However, this doesn't
happen when it's not possible to flatten the subquery, for example because
it contains GROUP BY, LIMIT, etc.  Allowing the subquery to compute useless
output columns is often fairly harmless, but sometimes it has significant
performance cost: the unused output might be an expensive expression,
or it might be a Var from a relation that we could remove entirely (via
the join-removal logic) if only we realized that we didn't really need
that Var.  Situations like this are common when expanding views, so it
seems worth taking the trouble to detect and remove unused outputs.

Because the upper query's Var numbering for subquery references depends on
positions in the subquery targetlist, we don't want to renumber the items
we leave behind.  Instead, we can implement "removal" by replacing the
unwanted expressions with simple NULL constants.  This wastes a few cycles
at runtime, but not enough to justify more work in the planner.

以及补丁：a4c35ea1c2f05dd5b56739fcd0dc36a4870ea0c0
Improve parser's and planner's handling of set-returning functions.
补丁：9d4444a6fc011b7e2ca16386cf9f1c9f25c114e5
Preserve exposed type of subquery outputs when substituting NULLs.

关于 remove_unused_subquery_outputs 的修改